+++ ortools/constraint_solver/routing_index_manager.cc
@@ -21,7 +21,7 @@
 
 namespace operations_research {
 
-const int RoutingIndexManager::kUnassigned = -1;
+const int64 RoutingIndexManager::kUnassigned = -1;
 
 RoutingIndexManager::RoutingIndexManager(int num_nodes, int num_vehicles,
                                          NodeIndex depot)
@@ -77,7 +77,7 @@ void RoutingIndexManager::Initialize(
   node_to_index_.resize(num_nodes_, kUnassigned);
   vehicle_to_start_.resize(num_vehicles_);
   vehicle_to_end_.resize(num_vehicles_);
-  int index = 0;
+  int64 index = 0;
   for (NodeIndex i = kZeroNode; i < num_nodes_; ++i) {
     if (gtl::ContainsKey(starts, i) || !gtl::ContainsKey(ends, i)) {
       index_to_node_[index] = i;
@@ -90,7 +90,7 @@ void RoutingIndexManager::Initialize(
     const NodeIndex start = starts_ends[i].first;
     if (!gtl::ContainsKey(seen_starts, start)) {
       seen_starts.insert(start);
-      const int start_index = node_to_index_[start];
+      const int64 start_index = node_to_index_[start];
       vehicle_to_start_[i] = start_index;
       CHECK_NE(kUnassigned, start_index);
     } else {
@@ -110,7 +110,7 @@ void RoutingIndexManager::Initialize(
   // Logging model information.
   VLOG(1) << "Number of nodes: " << num_nodes_;
   VLOG(1) << "Number of vehicles: " << num_vehicles_;
-  for (int index = 0; index < index_to_node_.size(); ++index) {
+  for (int64 index = 0; index < index_to_node_.size(); ++index) {
     VLOG(2) << "Variable index " << index << " -> Node index "
             << index_to_node_[index];
   }
@@ -120,12 +120,12 @@ void RoutingIndexManager::Initialize(
   }
 }
 
-std::vector<int> RoutingIndexManager::NodesToIndices(
+std::vector<int64> RoutingIndexManager::NodesToIndices(
     const std::vector<NodeIndex>& nodes) const {
-  std::vector<int> indices;
+  std::vector<int64> indices;
   indices.reserve(nodes.size());
   for (const NodeIndex node : nodes) {
-    const int index = NodeToIndex(node);
+    const int64 index = NodeToIndex(node);
     CHECK_NE(kUnassigned, index);
     indices.push_back(index);
   }
